seperated out hover and screenshotstrip into purecomponents
